+++ CHANGELOG.md
@@ -3,9 +3,11 @@
 ## Unreleased
 
 ### Misc. Updates
-
+- **AWSAuthSDK**
+  - Upgrade Facebook SDK to 8.0.0 ([PR #3082](https://github.com/aws-amplify/aws-sdk-ios/pull/3082))
 - Model updates for the following services
   - Amazon Connect
+  - Amazon DynamoDB
   - Amazon EC2
   - Amazon Elastic Load Balancing
   - Amazon Pinpoint